chore: inline feature flag gate
